requirements.txt
pytesseract==0.2.7
Pillow==6.1.0
pyTelegramBotAPI==3.6.6
